ai-summary-bot/requirements.txt
aiogram==3.4.1
openai
aiohttp
aiosqlite
python-dotenv
